chore!: Remove deprecated typescript-node hook
