+++ index.js
@@ -219,7 +219,6 @@ var extensions = {
   },
   '.ts': [
     'ts-node/register',
-    'typescript-node/register',
     'typescript-register',
     'typescript-require',
     'sucrase/register/ts',
@@ -258,7 +257,6 @@ var extensions = {
   ],
   '.tsx': [
     'ts-node/register',
-    'typescript-node/register',
     'sucrase/register',
     {
       module: '@babel/register',
