+++ test/fixtures/ts/1/package.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
-    "typescript": "1.5.3",
-    "typescript-node": "^0.1.3"
+    "typescript": "1.4.1",
+    "typescript-register": "^1.1.0"
   }
 }
