+++ test/fixtures/ts/2/package.json
@@ -1,6 +1,5 @@
 {
   "dependencies": {
-    "typescript": "1.4.1",
-    "typescript-register": "^1.1.0"
+    "typescript-require": "^0.2.10"
   }
 }
