+++ test/fixtures/ts/4/.babelrc
