+++ test/fixtures/ts/4/component.ts
