+++ test/fixtures/ts/5/package.json
@@ -1,9 +1,6 @@
 {
   "dependencies": {
-    "@babel/core": "^7.2.2",
-    "@babel/preset-env": "^7.2.3",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.1.0",
-    "@babel/register": "^7.0.0"
+    "esbuild": "^0.14.29",
+    "esbuild-register": "^3.3.2"
   }
 }
