+++ test/fixtures/ts/5/test.ts
@@ -1,5 +1,13 @@
-import Component from "./component"
+function add(x: number, y: number): number {
+  return x + y;
+}
 
-var test: typeof Component = { ...Component }
-
-export default test
+export default {
+  data: {
+    trueKey: true as boolean,
+    falseKey: false as boolean,
+    subKey: {
+      subProp: add(0.5, 0.5),
+    },
+  },
+};
