+++ test/fixtures/ts/6/.swcrc
