+++ test/fixtures/ts/6/package.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
-    "esbuild": "^0.14.29",
-    "esbuild-register": "^3.3.2"
+    "@swc/core": "^1.2.110",
+    "@swc/register": "^0.1.7"
   }
 }
