--- test/fixtures/ts/7/test.ts
@@ -1,15 +1,0 @@
-var test = {
-  data: {
-    trueKey: true,
-    falseKey: false,
-    subKey: {
-      subProp: 1
-    }
-  }
-};
-
-var main = {
-  default: test
-};
-
-export = main;
