+++ test/fixtures/tsx/2/.babelrc
