+++ test/fixtures/tsx/2/package.json
@@ -1,5 +1,9 @@
 {
   "dependencies": {
-    "sucrase": "^3.12.1"
+    "@babel/core": "^7.2.2",
+    "@babel/preset-env": "^7.2.3",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-typescript": "^7.1.0",
+    "@babel/register": "^7.0.0"
   }
 }
