+++ test/fixtures/tsx/3/data.tsx
@@ -1,19 +1,13 @@
-var test: {
-  data: {
-    trueKey: boolean,
-    falseKey: boolean,
-    subKey: {
-      subProp: number
-    }
+export const data: {
+  trueKey: boolean,
+  falseKey: boolean,
+  subKey: {
+    subProp: number
   }
 } = {
-  data: {
-    trueKey: true,
-    falseKey: false,
-    subKey: {
-      subProp: 1
-    }
-  }
-}
-
-export default test
+  trueKey: true,
+  falseKey: false,
+  subKey: {
+    subProp: 1,
+  },
+};
