+++ test/fixtures/tsx/3/test.tsx
@@ -1,20 +1,11 @@
-import data from "./data"
+/** @jsx jsx */
 
-const React = {
-  createElement(Component: () => any) {
-    return Component()
-  }
+import { data } from "./data";
+
+const Component = (props: object) => ({ data: props });
+
+function jsx(element: typeof Component, props: object) {
+  return element(props);
 }
 
-// Test harmony arrow functions.
-const Component = () => {
-  var trueKey: boolean = true
-  var falseKey: boolean = false
-  var subKey = { subProp: 1 }
-
-  // Test harmony object short notation.
-  return { data: { trueKey, falseKey, subKey } }
-}
-
-// Test TSX syntax.
-export default <Component {...data} />
+export default <Component {...data} />;
