+++ test/fixtures/tsx/4/.swcrc
