--- test/fixtures/tsx/4/data.tsx
@@ -1,13 +1,0 @@
-export const data: {
-  trueKey: boolean,
-  falseKey: boolean,
-  subKey: {
-    subProp: number
-  }
-} = {
-  trueKey: true,
-  falseKey: false,
-  subKey: {
-    subProp: 1,
-  },
-};
