+++ test/fixtures/tsx/4/test.tsx
@@ -1,11 +1,18 @@
-/** @jsx jsx */
-
-import { data } from "./data";
-
-const Component = (props: object) => ({ data: props });
-
-function jsx(element: typeof Component, props: object) {
-  return element(props);
+const React = {
+  createElement(Component: () => any) {
+    return Component()
+  }
 }
 
-export default <Component {...data} />;
+// Test harmony arrow functions.
+const Component = () => {
+  var trueKey: boolean = true
+  var falseKey: boolean = false
+  var subKey = { subProp: 1 }
+
+  // Test harmony object short notation.
+  return { data: { trueKey, falseKey, subKey } }
+};
+
+// Test TSX syntax.
+export default <Component />
