--- test/fixtures/tsx/5/package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "@swc/core": "^1.2.110",
-    "@swc/register": "^0.1.7"
-  }
-}
